Moved some code into eirbot2014 module
