+++ modules/eirbot2014/include/device/eirbot2014/encoder.hpp
@@ -58,4 +58,28 @@ private:
   bool _inverse;
 };
 
+////////////////////////////////////////////////////////////////////
+
+template<typename T>
+inline Encoder<T>::Encoder(const char* name, T* addr)
+  : Device(name), _addr(addr), _inverse(false) {
+
+}
+
+template<typename T>
+inline s32 Encoder<T>::getValue(void) {
+  if(_inverse) {
+    return -(*_addr);
+  }
+  else {
+    return *_addr;
+  }
+}
+
+template<typename T>
+inline void Encoder<T>::inverse(void) {
+  _inverse = !_inverse;
+}
+
+
 #endif//ENCODER_HPP
